LPS-179567 Extract CommerceDiscountDetailsScreenNavigationCategory methods to a CommerceDiscountDetailsScreenNavigationEntry component and extends CommerceDiscountDetailsScreenNavigationCategory in it
