+++ modules/apps/commerce/commerce-pricing-web/src/main/java/com/liferay/commerce/pricing/web/internal/frontend/taglib/servlet/taglib/CommerceDiscountDetailsScreenNavigationCategory.java
@@ -14,36 +14,13 @@
 
 package com.liferay.commerce.pricing.web.internal.frontend.taglib.servlet.taglib;
 
-import com.liferay.commerce.currency.service.CommerceCurrencyLocalService;
-import com.liferay.commerce.discount.model.CommerceDiscount;
-import com.liferay.commerce.discount.rule.type.CommerceDiscountRuleTypeRegistry;
-import com.liferay.commerce.discount.service.CommerceDiscountRuleService;
-import com.liferay.commerce.discount.service.CommerceDiscountService;
-import com.liferay.commerce.discount.target.CommerceDiscountTargetRegistry;
-import com.liferay.commerce.percentage.PercentageFormatter;
 import com.liferay.commerce.pricing.web.internal.constants.CommerceDiscountScreenNavigationConstants;
-import com.liferay.commerce.pricing.web.internal.display.context.CommerceDiscountDisplayContext;
 import com.liferay.frontend.taglib.servlet.taglib.ScreenNavigationCategory;
-import com.liferay.frontend.taglib.servlet.taglib.ScreenNavigationEntry;
-import com.liferay.frontend.taglib.servlet.taglib.util.JSPRenderer;
 import com.liferay.portal.kernel.language.Language;
-import com.liferay.portal.kernel.log.Log;
-import com.liferay.portal.kernel.log.LogFactoryUtil;
-import com.liferay.portal.kernel.model.User;
-import com.liferay.portal.kernel.security.permission.ActionKeys;
-import com.liferay.portal.kernel.security.permission.PermissionThreadLocal;
-import com.liferay.portal.kernel.security.permission.resource.ModelResourcePermission;
-import com.liferay.portal.kernel.util.Portal;
 import com.liferay.portal.kernel.util.ResourceBundleUtil;
-import com.liferay.portal.kernel.util.WebKeys;
-
-import java.io.IOException;
 
 import java.util.Locale;
 import java.util.ResourceBundle;
-
-import javax.servlet.http.HttpServletRequest;
-import javax.servlet.http.HttpServletResponse;
 
 import org.osgi.service.component.annotations.Component;
 import org.osgi.service.component.annotations.Reference;
@@ -52,23 +29,14 @@ import org.osgi.service.component.annotations.Reference;
  * @author Riccardo Alberti
  */
 @Component(
-	property = {
-		"screen.navigation.category.order:Integer=10",
-		"screen.navigation.entry.order:Integer=10"
-	},
-	service = {ScreenNavigationCategory.class, ScreenNavigationEntry.class}
+	property = "screen.navigation.category.order:Integer=10",
+	service = ScreenNavigationCategory.class
 )
 public class CommerceDiscountDetailsScreenNavigationCategory
-	implements ScreenNavigationCategory,
-			   ScreenNavigationEntry<CommerceDiscount> {
+	implements ScreenNavigationCategory {
 
 	@Override
 	public String getCategoryKey() {
-		return CommerceDiscountScreenNavigationConstants.CATEGORY_KEY_DETAILS;
-	}
-
-	@Override
-	public String getEntryKey() {
 		return CommerceDiscountScreenNavigationConstants.CATEGORY_KEY_DETAILS;
 	}
 
@@ -77,7 +45,7 @@ public class CommerceDiscountDetailsScreenNavigationCategory
 		ResourceBundle resourceBundle = ResourceBundleUtil.getBundle(
 			"content.Language", locale, getClass());
 
-		return _language.get(
+		return language.get(
 			resourceBundle,
 			CommerceDiscountScreenNavigationConstants.CATEGORY_KEY_DETAILS);
 	}
@@ -88,85 +56,7 @@ public class CommerceDiscountDetailsScreenNavigationCategory
 			SCREEN_NAVIGATION_KEY_DISCOUNT_GENERAL;
 	}
 
-	@Override
-	public boolean isVisible(User user, CommerceDiscount commerceDiscount) {
-		if (commerceDiscount == null) {
-			return false;
-		}
-
-		boolean hasPermission = false;
-
-		try {
-			hasPermission = _commerceDiscountModelResourcePermission.contains(
-				PermissionThreadLocal.getPermissionChecker(),
-				commerceDiscount.getCommerceDiscountId(), ActionKeys.UPDATE);
-		}
-		catch (Exception exception) {
-			if (_log.isDebugEnabled()) {
-				_log.debug(exception);
-			}
-		}
-
-		return hasPermission;
-	}
-
-	@Override
-	public void render(
-			HttpServletRequest httpServletRequest,
-			HttpServletResponse httpServletResponse)
-		throws IOException {
-
-		CommerceDiscountDisplayContext commerceDiscountDisplayContext =
-			new CommerceDiscountDisplayContext(
-				_commerceCurrencyLocalService,
-				_commerceDiscountModelResourcePermission,
-				_commerceDiscountService, _commerceDiscountRuleService,
-				_commerceDiscountRuleTypeRegistry,
-				_commerceDiscountTargetRegistry, _percentageFormatter,
-				httpServletRequest, _portal);
-
-		httpServletRequest.setAttribute(
-			WebKeys.PORTLET_DISPLAY_CONTEXT, commerceDiscountDisplayContext);
-
-		_jspRenderer.renderJSP(
-			httpServletRequest, httpServletResponse,
-			"/commerce_discounts/details.jsp");
-	}
-
-	private static final Log _log = LogFactoryUtil.getLog(
-		CommerceDiscountDetailsScreenNavigationCategory.class);
-
 	@Reference
-	private CommerceCurrencyLocalService _commerceCurrencyLocalService;
-
-	@Reference(
-		target = "(model.class.name=com.liferay.commerce.discount.model.CommerceDiscount)"
-	)
-	private ModelResourcePermission<CommerceDiscount>
-		_commerceDiscountModelResourcePermission;
-
-	@Reference
-	private CommerceDiscountRuleService _commerceDiscountRuleService;
-
-	@Reference
-	private CommerceDiscountRuleTypeRegistry _commerceDiscountRuleTypeRegistry;
-
-	@Reference
-	private CommerceDiscountService _commerceDiscountService;
-
-	@Reference
-	private CommerceDiscountTargetRegistry _commerceDiscountTargetRegistry;
-
-	@Reference
-	private JSPRenderer _jspRenderer;
-
-	@Reference
-	private Language _language;
-
-	@Reference
-	private PercentageFormatter _percentageFormatter;
-
-	@Reference
-	private Portal _portal;
+	protected Language language;
 
 }
